major, major change to permit passing a model or models directly and defining model classes accordingly.
also a bug fix in that if you declare a before_foo or after_foo step, and you don't declare that step, the before and/or after still get run.

finally, a change that data is always passed as locals, even if one of the locals is "model." This corrects a major, major bug where unobtrusive renderers weren't getting the default data stream.